dictmaker/pyproject.toml
[project]
name = "dictmaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "beautifulsoup4>=4.14.3",
    "certifi>=2026.1.4",
    "charset-normalizer==3.4.4",
    "colorama==0.4.6",
    "comtypes>=1.4.14",
    "deprecated==1.3.1",
    "greenlet==3.3.0",
    "html2text==2025.4.15",
    "idna==3.11",
    "iniconfig==2.3.0",
    "jaconv==0.4.0",
    "packaging==25.0",
    "pluggy==1.6.0",
    "pydantic>=2.12.5",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "pykakasi==2.3.0",
    "pyqt5==5.15.11",
    "pyqt5-qt5==5.15.2",
    "pyqt5-sip==12.17.2",
    "pytest==9.0.2",
    "python-dotenv==1.2.1",
    "pywin32==311",
    "pywinauto==0.6.9",
    "requests==2.32.5",
    "six==1.17.0",
    "soupsieve>=2.8.1",
    "sqlalchemy==2.0.45",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3>=2.6.2",
    "wrapt==2.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]

[tool.uv]
cache-dir = "./.uv_cache"
